Postfix generator Merged into Parser
